begin replacing report with report browser
