+++ client/src/views/Researcher/ReportBrowser.jsx
@@ -1,0 +1,17 @@
+import NavBar from '../../components/NavBar/NavBar';
+import RouteButton from '../../components/RouteButton/RouteButton';
+import React, { useState } from 'react';
+import { Link } from 'react-router-dom';
+import './Report.less';
+
+export default function Report(props) {
+  return (
+    <div className='container nav-padding'>
+      <NavBar />
+      <div id='main-header'>Welcome Researcher!</div>
+      <h1 id='report-subheader'>Reports</h1>
+      <div id='button-container'>
+      </div>
+    </div>
+  );
+}
